pyproject.toml
[project]
authors = [
  { name = "Gustave Coulombe", email = "magikgus@gmail.com" },
  { name = "Yannick Lapointe", email = "yannicklapointe@proton.me" },
]
name = "graphinglib"
version = "1.6.0"
description = "An object oriented wrapper combining the functionalities of Matplotlib and Scipy"
license = "MIT"
readme = "README.md"
requires-python = ">=3.12,<3.15"
dependencies = [
  "numpy>=1.24.2,<2.0.0",
  "scipy>=1.10.1,<2.0.0",
  "matplotlib>=3.7.1,<4.0.0",
  "PyYAML>=6.0,<7.0",
  "platformdirs>=3.11.0,<4.0.0",
  "setuptools-scm>=8.0.4,<9.0.0",
  "shapely>=2.0.3,<3.0.0",
  "pyperclip>=1.9.0,<2.0.0",
]

[project.optional-dependencies]
astro = [
  "astropy>=7.0.0,<8.0.0",
]

[dependency-groups]
dev = [
  "ipykernel>=6.23.0,<7.0.0",
  "sphinx-favicon>=1.0.1,<2.0.0",
  "sphinx-copybutton>=0.5.2,<0.6.0",
  "sphinx-design>=0.5.0,<0.6.0",
  "pydata-sphinx-theme>=0.14.1,<0.15.0",
  "numpydoc>=1.7.0,<2.0.0",
  "pygithub>=2.3.0,<3.0.0",
  "astropy>=7.0.0,<8.0.0",
  "ruff>=0.15.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true
